Refactored all files with ReSharper to make variable name consistant
CYPHER BREAKING CHANGE:
Removed the ability to pass a cypher query string to the post method,
Cypher querys must be built using the Cypher object.  This change was made
to support the better type handeling of returned object (see below)

Ehnanced the retun type handeling of cypher querys. The returned table
from a cypher post will have the correct data types based on the building
of query instead of the returned json.  this prevents the possible error
of not knowing the return type for a column for optional parameters and
the first row containing a null value.